Create Store Database
I created a decently sized store database for a mom and pop store. The data in my database is fairly large, so this is just the backbone of my database. I created the database naming it Storedb. Secondly, I added 4 main tables consisting of Customers, Employees, Products and Transactions. There are 50 customers with their respective values in my Customer table, 20 Employees and their respective values in my Employee Table, 141 different products and their respective values in the Products table, and 50 transactions and their respective values in the Transactions table.