Create Topmovies Database and Stats
For this file, I performed a data dig on the topmovies database. Firstly I executed a MAX, AVG, and MIN function calculation on the data for the values of Domestic, Year and Overseas. Secondly, I executed a HAVING function calculation on the overall titles in topmovies whose studios were P/DW. Thirdly, I used the CASE function calculation to create my own column named Film_Genre where each film title has their corresponding list of genres.
Last but not least, I used an AND and an OR calculation to narrow my search for films that I grew up watching.